+++ src/app/admin/layout.tsx
@@ -3,14 +3,21 @@ import { Montserrat } from "next/font/google";
 import Sidebar from "@components/Sidebar";
 import style from "@styles/admin/layout.module.css";
 import TabBar from "../components/TabBar";
+import { getUserDbData } from "app/lib/authentication";
+import { redirect } from "next/navigation";
 type Props = {
   children: ReactNode;
 };
 
 const montserrat = Montserrat({ subsets: ["latin"], weight: ["300"] });
-const Layout = (props: Props) => {
+const Layout = async (props: Props) => {
  
+    // get user if possible
+    const user = await getUserDbData()
 
+    if (user?.role != "admin"){
+        redirect("/")
+    }
 
   return (
     <>
